FIN-13 Moving hooks folder to utils
