+++ src/src_V2/components/tooltip/TooltipText.jsx
@@ -1,7 +1,7 @@
 import { useRef } from 'react';
 import { Box, Tooltip } from '@mui/material';
 
-import { useHoverTooltip } from '../../hooks';
+import { useHoverTooltip } from '../../utils/hooks';
 import styles from './TooltipText.module.css';
 
 export default function TooltipText({ text, maxWidth }) {
@@ -10,8 +10,12 @@ export default function TooltipText({ text, maxWidth }) {
   const hoverStatus = useHoverTooltip(text, textElementRef);
 
   return (
-    <Tooltip title={text} interactive="true" disableHoverListener={!hoverStatus}>
-      <Box ref={textElementRef} className={styles.tooltip} sx={{ maxWidth: maxWidth || '100px' }}>
+    <Tooltip title={text} interactive='true' disableHoverListener={!hoverStatus}>
+      <Box
+        ref={textElementRef}
+        className={styles.tooltip}
+        sx={{ maxWidth: maxWidth || '100px' }}
+      >
         {text}
       </Box>
     </Tooltip>
