+++ src/src_V2/modules/manager/reporting/index.jsx
@@ -1,6 +1,6 @@
 import { Table } from '../../../components';
 import tourReportTableDataGenerator from './utils/table-data-generator';
-import { useTableDataGenerator } from '../../../hooks';
+import { useTableDataGenerator } from '../../../utils/hooks';
 import { Container } from '../../../components';
 
 import { Typography } from '@mui/material';
